+++ examples/springboot/src/main/java/io/hawt/example/spring/boot/SampleSpringBootService.java
@@ -4,6 +4,7 @@ import io.hawt.config.ConfigFacade;
 import io.hawt.springboot.HawtPlugin;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
 import org.springframework.context.annotation.Bean;
 
 @SpringBootApplication